Finished Arc Curve
The docs aren't exactly copied from SketchUp Ruby API.
Instead of saying, "The normal method retrieves a Vector3d that is
perpendicular to the plane of the arc," we say, "Get arc normal, the
vector that is perpendicular to the plane of the arc." I think that we
don't need to repeat method's name in description because this
description is placed in the methods's description section, which tells
us that this description belongs to this method. We also replace
Vector3d in description with a simple word vector because the type of
vector is specified in return value.